Moved mailer to utils
