+++ api/src/routes/purchase.js
@@ -20,8 +20,9 @@ import * as fs from "fs/promises";
 import jimp from "jimp";
 
 import Models from "../models/index.js";
-import { mailer, ACCESS_LEVEL, cleanUTF8 } from "../common_items.js";
+import { ACCESS_LEVEL, cleanUTF8 } from "../common_items.js";
 import { logger } from "../utils/logging.js";
+import { mailer } from "../utils/mailer.js";
 import { committee_id_to_display } from "../db_loaded_items.js";
 
 // filter uploaded files based on type
